+++ packages/frontend/src/components/message/Message.tsx
@@ -301,6 +301,12 @@ function buildContextMenu(
       label: tx('react'),
       action: handleReactClick,
     },
+    // Save Message
+    {
+      label: tx('save'),
+      action: () =>
+        BackendRemote.rpc.saveMsgs(selectedAccountId(), [message.id]),
+    },
     // copy link
     link !== '' &&
       isLink && {
@@ -414,7 +420,6 @@ export default function Message(props: {
       >
     ) => {
       event.preventDefault() // prevent default runtime context menu from opening
-
       const chat = await BackendRemote.rpc.getFullChatById(
         accountId,
         message.chatId
@@ -634,6 +639,12 @@ export default function Message(props: {
     MessageTagName = 'button'
   }
 
+  // Check if the message is saved or has a saved message
+  // in both cases we display the bookmark icon
+  const isOrHasSavedMessage = message.originalMsgId
+    ? true
+    : !!message.savedMessageId
+
   let content
   if (message.viewType === 'VideochatInvitation') {
     return (
@@ -681,6 +692,7 @@ export default function Message(props: {
             hasLocation={hasLocation}
             timestamp={message.timestamp * 1000}
             padlock={message.showPadlock}
+            isSavedMessage={isOrHasSavedMessage}
             onClickError={openMessageInfo.bind(null, openDialog, message)}
             viewType={'VideochatInvitation'}
             tabindexForInteractiveContents={tabindexForInteractiveContents}
@@ -731,7 +743,9 @@ export default function Message(props: {
 
   /** Whether to show author name and avatar */
   const showAuthor =
-    conversationType.hasMultipleParticipants || message?.overrideSenderName
+    conversationType.hasMultipleParticipants ||
+    message?.overrideSenderName ||
+    message?.originalMsgId
 
   const hasText = text !== null && text !== ''
   const fileMime = (!isSetupmessage && message.fileMime) || null
@@ -862,6 +876,7 @@ export default function Message(props: {
               hasLocation={hasLocation}
               timestamp={message.timestamp * 1000}
               padlock={message.showPadlock}
+              isSavedMessage={isOrHasSavedMessage}
               onClickError={openMessageInfo.bind(null, openDialog, message)}
               viewType={message.viewType}
               tabindexForInteractiveContents={tabindexForInteractiveContents}